# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdaptRead API",
    "description": "Contextual-bandit presentation service backed by simulated-reader experiments.",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies": {
      "get": {
        "summary": "Policies",
        "operationId": "policies_v1_policies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Policies V1 Policies Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/simulator": {
      "get": {
        "summary": "Simulator",
        "operationId": "simulator_v1_simulator_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulator V1 Simulator Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Session V1 Sessions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/step": {
      "post": {
        "summary": "Step Session",
        "operationId": "step_session_v1_sessions__session_id__step_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Step Session V1 Sessions  Session Id  Step Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/trace": {
      "get": {
        "summary": "Session Trace",
        "operationId": "session_trace_v1_sessions__session_id__trace_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Session Trace V1 Sessions  Session Id  Trace Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConstraintInput": {
        "properties": {
          "min_font_size": {
            "type": "integer",
            "title": "Min Font Size",
            "default": 16
          },
          "tts_allowed": {
            "type": "boolean",
            "title": "Tts Allowed",
            "default": true
          },
          "max_changed_dimensions": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Max Changed Dimensions",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ConstraintInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SessionInput": {
        "properties": {
          "constraints": {
            "$ref": "#/components/schemas/ConstraintInput",
            "default": {
              "min_font_size": 16,
              "tts_allowed": true,
              "max_changed_dimensions": 2
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SessionInput"
      },
      "TelemetryInput": {
        "properties": {
          "observed_wpm": {
            "type": "number",
            "maximum": 1500.0,
            "minimum": 0.0,
            "title": "Observed Wpm"
          },
          "regression_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Regression Rate"
          },
          "pause_ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Pause Ratio"
          },
          "probe_correct": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probe Correct"
          },
          "segment_difficulty": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Segment Difficulty"
          },
          "segment_length": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Segment Length"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observed_wpm",
          "regression_rate",
          "pause_ratio",
          "probe_correct",
          "segment_difficulty",
          "segment_length"
        ],
        "title": "TelemetryInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}